update integration & derivation operations

+++ duino.ino/derivation.h
@@ -1,5 +1,5 @@
 // period is in milliseconds
-float derivation(unsigned long period, long err_actual, long err_prev){
+float derivation(int period, int err_actual, int err_prev){
   int err = err_actual - err_prev;
   return static_cast<float>(err)/period;
 }
